test: split test into storage, distribution-{array,lru}
